pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
discord-py-slash-command = "4.2.1"
requests = "^2.31.0"
sqlalchemy = "^2.0.28"
flask = "^3.0.2"
interaction = "^1.3"
awsiot = "^0.1.3"
discord-js = "1.0.7"
discord-py = "2.3.2"
xds-protos = "^1.62.1"
openai = "^1.14.2"
akinator = "^1.1.1"
pytz = "^2024.1"
cryptography = "^42.0.5"
pyjokes = "^0.6.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"